Loss of precision, Time being returned as a float rather than a double.

+++ src/CModelList.h
@@ -39,7 +39,7 @@ public:
 	void GetAllParameters(float * params, int n_params);
 	void GetFreeParameters(float * params, int n_params);
 	CModel * GetModel(int i) { return mList[i]; };
-	float GetTime() { return mTime; };
+	double GetTime() { return mTime; };
 
 	vector< pair<eModels, string> > GetList_AllModels(void);
